moonshot/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
cssselect==1.1.0
Django==4.1.1
django-crispy-forms==1.14.0
django-environ==0.9.0
django-modeltranslation==0.18.4
django-rosetta==0.9.8
django-summernote==0.8.20.0
idna==3.4
lxml==4.9.1
markdown2==2.4.3
polib==1.1.1
pyquery==1.4.3
readtime==2.0.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
webencodings==0.5.1
